+++ libft/srcs/ft_count_elt.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_count.c                                         :+:      :+:    :+:   */
+/*   ft_count_elt.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: abarot <abarot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/28 17:55:46 by abarot            #+#    #+#             */
-/*   Updated: 2020/09/03 13:57:29 by abarot           ###   ########.fr       */
+/*   Updated: 2020/09/10 09:36:44 by abarot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_count_elt(char *str, char *elt)
+size_t	ft_count_elt(char *str, char *elt)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	if (!str || !elt)
